apps/backend/requirements.txt
# ============================================================
# VegasDelRio Backend - Dependencias Python
# ============================================================

# --- Framework Web ---
fastapi==0.115.*
uvicorn[standard]==0.34.*

# --- Base de Datos (ORM + Driver Async) ---
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
alembic==1.14.*
psycopg2-binary==2.9.*          # Requerido por Alembic (migraciones síncronas)

# --- Validación y Configuración ---
pydantic==2.10.*
pydantic-settings==2.7.*
python-dotenv==1.0.*

# --- Autenticación ---
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*

# --- Celery (Tareas Asíncronas) ---
celery[redis]==5.4.*
redis==5.2.*

# --- Generación de PDFs ---
weasyprint==63.*
jinja2==3.1.*

# --- Envío de Correos ---
resend==2.5.*                   # API de email transaccional (alternativa)
google-auth==2.38.*             # Credenciales Google OAuth
google-auth-oauthlib==1.2.*     # Flujo OAuth para Google
google-api-python-client==2.163.*  # Gmail API client

# --- Utilidades ---
python-multipart==0.0.*         # Requerido por FastAPI para formularios
httpx==0.28.*                   # Cliente HTTP async

# --- Testing ---
pytest==8.3.*
pytest-asyncio==0.25.*
pytest-cov==6.1.*
aiosqlite==0.21.*               # SQLite async para tests in-memory
